# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: Contract-first API for browsing products and creating orders.
security: []
servers:
  - url: http://localhost:3000
    description: Local development server
tags:
  - name: Products
    description: Marketplace product catalog
  - name: Orders
    description: Customer orders
paths:
  /products:
    get:
      tags: [Products]
      operationId: listProducts
      summary: List products
      description: Returns products using cursor-based pagination.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of products.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
        "400":
          $ref: "#/components/responses/BadRequest"
  /products/{productId}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [Products]
      operationId: getProduct
      summary: Get a product
      description: Returns one product by its identifier.
      responses:
        "200":
          description: The requested product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders:
    get:
      tags: [Orders]
      operationId: listOrders
      summary: List orders
      description: Returns orders using cursor-based pagination.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Create an order
      description: Creates an order and calculates all monetary values in integer cents.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Reusing the same key with the same request body returns the same 201 response; reusing it with a different body returns 422.
          schema:
            type: string
            minLength: 1
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: The order was created, or the original successful response was replayed.
          headers:
            Location:
              description: Relative URL of the created order.
              schema:
                type: string
            Idempotency-Replay:
              description: Present and set to true when a stored response is replayed.
              schema:
                type: string
                enum: ["true"]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /orders/{orderId}:
    parameters:
      - $ref: "#/components/parameters/OrderId"
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Get an order
      description: Returns one order by its identifier.
      responses:
        "200":
          description: The requested order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of resources to return.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      description: Opaque token returned as next_cursor by the previous page; clients must not inspect or construct it.
      required: false
      schema:
        type: string
        minLength: 1
    ProductId:
      name: productId
      in: path
      required: true
      description: Product identifier.
      schema:
        type: integer
        minimum: 1
    OrderId:
      name: orderId
      in: path
      required: true
      description: Order identifier.
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: The request does not conform to the API contract.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: The request is valid but conflicts with an earlier idempotent request.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Product:
      type: object
      additionalProperties: false
      required: [id, name, price_cents]
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
        price_cents:
          type: integer
          minimum: 0
          description: Product price in the smallest currency unit.
    OrderItemInput:
      type: object
      additionalProperties: false
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 100
    OrderItem:
      type: object
      additionalProperties: false
      required: [product_id, quantity, unit_price_cents, line_total_cents]
      properties:
        product_id:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 100
        unit_price_cents:
          type: integer
          minimum: 0
        line_total_cents:
          type: integer
          minimum: 0
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/OrderItemInput"
    Order:
      type: object
      additionalProperties: false
      required: [id, items, total_cents, created_at]
      properties:
        id:
          type: integer
          minimum: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        total_cents:
          type: integer
          minimum: 0
          description: Order total in the smallest currency unit.
        created_at:
          type: string
          format: date-time
    ProductPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page, or null when no pages remain.
    OrderPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page, or null when no pages remain.
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the error type.
        title:
          type: string
          description: Short, human-readable error title.
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: HTTP status code.
        detail:
          type: string
          description: Explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference identifying this error occurrence.